pyproject.toml
[project]
name = "shelfscope"
version = "0.1.0"
description = "FastAPI app for retail shelf detection and visual grouping"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "transformers>=4.48.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy<2",
    "pillow>=10.0.0",
    "opencv-python>=4.9.0",
    "huggingface-hub>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
]

[tool.uv]
package = false

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]
torchvision = [{ index = "pytorch-cu126" }]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
